use channel instead of mutex to queue update in a store
