+++ core/src/jsMain/kotlin/dev/fritz2/core/handlers.kt
@@ -1,5 +1,6 @@
 package dev.fritz2.core
 
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.FlowCollector
 import kotlinx.coroutines.flow.MutableSharedFlow
@@ -33,7 +34,7 @@ fun <T> flowOnceOf(value: T) = OnlyOnceFlow(value)
  * @property process function describing how this handler collects a [Flow] when called
  */
 interface Handler<A> {
-    val process: (Flow<A>) -> Unit
+    val process: (Flow<A>, Job) -> Unit
 }
 
 /**
@@ -42,7 +43,7 @@ interface Handler<A> {
  *
  * @param process defines how to handle the values of the connected [Flow]
  */
-class SimpleHandler<A>(override inline val process: (Flow<A>) -> Unit) : Handler<A>
+value class SimpleHandler<A>(override inline val process: (Flow<A>, Job) -> Unit) : Handler<A>
 
 /**
  * An [EmittingHandler] is a special [Handler] that constitutes a new [Flow] by itself. You can emit values to this [Flow] from your code
@@ -52,11 +53,11 @@ class SimpleHandler<A>(override inline val process: (Flow<A>) -> Unit) : Handler
  * @property process function defining how this [Handler] collects a [Flow] when connected using [handledBy]
  */
 class EmittingHandler<A, E>(
-    inline val collectWithChannel: (Flow<A>, FlowCollector<E>) -> Unit,
+    inline val collectWithChannel: (Flow<A>, FlowCollector<E>, Job) -> Unit,
     private val flow: MutableSharedFlow<E> = MutableSharedFlow()
 ) : Handler<A>, Flow<E> by flow {
 
-    override val process: (Flow<A>) -> Unit = { upstream ->
-        collectWithChannel(upstream, flow)
+    override val process: (Flow<A>, Job) -> Unit = { upstream, job ->
+        collectWithChannel(upstream, flow, job)
     }
 }
