+++ core/src/jsMain/kotlin/dev/fritz2/core/job.kt
@@ -39,7 +39,7 @@ interface WithJob {
      * @param handler [Handler] that will be called for each action/event on the [Flow]
      * @receiver [Flow] of action/events to bind to a [Handler]
      */
-    infix fun <A> Flow<A>.handledBy(handler: Handler<A>) = handler.process(this)
+    infix fun <A> Flow<A>.handledBy(handler: Handler<A>) = handler.process(this, job)
 
     /**
      * Connects a [Flow] to a suspendable [execute] function.
@@ -58,7 +58,7 @@ interface WithJob {
      * @param handler that will handle the fired [Event]
      */
     infix fun <E : Event> Flow<E>.handledBy(handler: Handler<Unit>) =
-        handler.process(this.map { })
+        handler.process(this.map { }, job)
 
     /**
      * Connects a [Flow] to a suspendable [execute] function.
@@ -74,12 +74,12 @@ interface WithJob {
      *
      * @param data parameter forwarded to the handler
      */
-    operator fun <A> Handler<A>.invoke(data: A) = this.process(flowOnceOf(data))
+    operator fun <A> Handler<A>.invoke(data: A) = this.process(flowOnceOf(data), job)
 
     /**
      * Calls this handler exactly once.
      */
-    operator fun Handler<Unit>.invoke() = this.process(flowOnceOf(Unit))
+    operator fun Handler<Unit>.invoke() = this.process(flowOnceOf(Unit), job)
 
 }
 
@@ -89,7 +89,7 @@ interface WithJob {
  * @param handler [Handler] that will be called for each action/event on the [Flow]
  * @receiver [Flow] of action/events to bind to an [Handler]
  */
-infix fun <A> Flow<A>.handledBy(handler: Handler<A>) = handler.process(this)
+infix fun <A> Flow<A>.handledBy(handler: Handler<A>) = handler.process(this, Job())
 
 /**
  * Connects a [Flow] to a suspendable [execute] function.
