+++ core/src/jsTest/kotlin/dev/fritz2/core/job.kt
@@ -1,11 +1,13 @@
 package dev.fritz2.core
 
-import kotlin.test.*
 import dev.fritz2.runTest
 import kotlinx.browser.document
 import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.map
 import org.w3c.dom.HTMLButtonElement
+import kotlin.test.*
 
 
 class AdHocHandlerTests {
@@ -115,5 +117,48 @@ class AdHocHandlerTests {
             document.getElementById(idWorkState)?.textContent,
             "Handler has finished and set state to `true`"
         )
+    }
+
+    @Test
+    fun testAdHocErrorHandling() = runTest {
+
+        val valueId = Id.next()
+        fun getValue() = document.getElementById(valueId)?.textContent
+
+        val startValue = "start"
+
+        val store = object : RootStore<String>(startValue) {
+        }
+
+        render {
+            div {
+                span(id = valueId) { store.data.renderText() }
+            }
+        }
+
+        val updates = MutableStateFlow(startValue)
+        updates handledBy store.update
+
+        suspend fun checkUpdate(msg: String) {
+            val valueAfterSuccessfullUpdate = Id.next()
+            updates.value = valueAfterSuccessfullUpdate
+            delay(150)
+            assertEquals(valueAfterSuccessfullUpdate, getValue(), msg)
+        }
+
+        checkUpdate("store not updating after start")
+
+        val adHocException = "adHoc exception"
+        flowOf(Unit) handledBy {
+            throw Exception(adHocException)
+        }
+        delay(150)
+        checkUpdate("store not updating after adhoc handler ")
+
+        val adHocIntermediateException = "adHoc intermediate exception"
+        flowOf(2).map {throw Exception(adHocIntermediateException)} handledBy {
+        }
+        delay(150)
+        checkUpdate("store not updating after intermediate exception before adhoc handler ")
     }
 }
